backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
Pillow>=10.0
django-cloudinary-storage>=0.3.0
cloudinary>=1.36
celery>=5.3
redis>=5.0
python-dotenv>=1.0
pytesseract>=0.3.10
pdf2image>=1.16
qrcode[pil]>=7.4
django-filter>=24.0
google-auth>=2.29
